Feature: Converts application into a Node.js App Removes: DOM dependency. Build: Adds browserify
