+++ examples/add.multiple.markers.html
@@ -12,7 +12,7 @@
   }
 </style>
 <body>
-<script src="../lib/index.js"></script>
+<script src="../dist/gmplus.js"></script>
 <script>
   var map = new GMP({
     id: 'myMap',
