+++ examples/crossfilter.html
@@ -17,7 +17,7 @@
   }
 </style>
 <body>
-<script src="../lib/index.js"></script>
+<script src="../dist/gmplus.js"></script>
 <script src="../node_modules/crossfilter/crossfilter.js"></script>
 <script src="../node_modules/jquery/dist/jquery.js"></script>
 <script>
