+++ lib/gmplus/map.js
@@ -1,5 +1,3 @@
-/* gmplus.js 0.1.0 MIT License. 2015 Yago Ferrer <yago.ferrer@gmail.com> */
-(function e(t,n,r){function s(o,u){if(!n[o]){if(!t[o]){var a=typeof require=="function"&&require;if(!u&&a)return a(o,!0);if(i)return i(o,!0);var f=new Error("Cannot find module '"+o+"'");throw f.code="MODULE_NOT_FOUND",f}var l=n[o]={exports:{}};t[o][0].call(l.exports,function(e){var n=t[o][1][e];return s(n?n:e)},l,l.exports,e,t,n,r)}return n[o].exports}var i=typeof require=="function"&&require;for(var o=0;o<r.length;o++)s(r[o]);return s})({1:[function(require,module,exports){
 var utils = require('../gmplus/utils');
 
 'use strict';
@@ -324,23 +322,3 @@ module.exports = function(global) {
   return GMP;
 
 };
-
-},{"../gmplus/utils":2}],2:[function(require,module,exports){
-function clone(o) {
-  var out, v, key;
-  out = Array.isArray(o) ? [] : {};
-  for (key in o) {
-    v = o[key];
-    out[key] = (typeof v === "object") ? clone(v) : v;
-  }
-  return out;
-}
-
-module.exports = {
-  clone: clone
-}
-
-},{}],3:[function(require,module,exports){
-window.GMP = require('./gmplus/map')(window);
-
-},{"./gmplus/map":1}]},{},[3]);
