+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "prestart": "npm install",
     "start": "browser-sync start --server --index examples/index.html",
-    "test": "mocha-phantomjs test/runner.html",
+    "test": "mocha -r ./test/bootstrap.js test/unit",
+    "predist": "browserify lib/index.js > dist/gmplus.js",
     "dist": "node ./scripts/dist.js",
     "pretest": "jshint lib/index.js"
   },
@@ -28,15 +29,14 @@
   "devDependencies": {
     "bower": "^1.3.12",
     "browser-sync": "^1.9.1",
+    "browserify": "^8.1.1",
     "chai": "^1.10.0",
     "crossfilter": "^1.3.11",
     "install": "^0.1.8",
     "jquery": "^2.1.3",
     "jshint": "^2.6.0",
     "mocha": "^2.1.0",
-    "mocha-phantomjs": "^3.5.3",
     "npm": "^2.2.0",
-    "phantomjs": "^1.9.13",
     "uglify-js": "^2.4.16"
   }
 }
