+++ scripts/dist.js
@@ -1,6 +1,6 @@
 var UglifyJS = require('uglify-js');
 var fs = require('fs');
-var file = 'lib/index.js';
+var file = 'dist/gmplus.js';
 var result = UglifyJS.minify(file, {
   mangle: true,
   compress: {
@@ -18,6 +18,7 @@ var result = UglifyJS.minify(file, {
 var pjson = require('../package.json');
 var credits = '/* gmplus.js ' + pjson.version + ' MIT License. ' + new Date().getFullYear() + ' Yago Ferrer <yago.ferrer@gmail.com> */\n';
 fs.writeFileSync('dist/gmplus.min.js', credits + result.code);
+
 var data = fs.readFileSync(file, 'utf-8');
 fs.writeFileSync('dist/gmplus.js', credits + data, 'utf-8');
