+++ test/mocks/window.js
@@ -1,0 +1,5 @@
+module.exports = {
+  document: {
+    getElementById: function() {}
+  }
+};
