--- test/runner.html
@@ -1,25 +1,0 @@
-<html>
-<head>
-  <meta charset="utf-8">
-  <link rel="stylesheet" href="../node_modules/mocha/mocha.css" />
-</head>
-<body>
-<div id="mocha"></div>
-
-<script src="../node_modules/mocha/mocha.js"></script>
-<script src="../node_modules/chai/chai.js"></script>
-<script>
-  mocha.ui('bdd');
-  mocha.reporter('html');
-  expect = chai.expect;
-</script>
-<script src="../lib/index.js"></script>
-<script src="mocks/google.maps.js"></script>
-<script src="unit/test.js"></script>
-<script>
-  if (window.mochaPhantomJS) { mochaPhantomJS.run(); }
-  else { mocha.run(); }
-</script>
-<div id="myMap"></div>
-</body>
-</html>
